+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * @iov/crypto: Add `Secp256k1.recoverPubkey` to recover pubkey from signature
   and message
+* @iov/ethereum: Implement pubkey recovery when parsing transaction in search
+  by transaction ID.
 
 ## 0.12.1